# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProcuraX API",
    "description": "\nProcuraX — procurement & spend transformation platform.\n\nPurchase-to-pay workflow with deterministic policy enforcement, multi-tenant isolation\n(application scoping + PostgreSQL Row-Level Security) and a complete audit trail.\n\n**Design principle:** AI recommends, rules enforce, humans decide on high-risk actions.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/vendors": {
      "get": {
        "tags": [
          "vendors"
        ],
        "summary": "List Vendors",
        "operationId": "list_vendors_api_v1_vendors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SpendCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VendorStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VendorRiskLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vendors"
        ],
        "summary": "Create Vendor",
        "operationId": "create_vendor_api_v1_vendors_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vendors/{vendor_id}": {
      "get": {
        "tags": [
          "vendors"
        ],
        "summary": "Get Vendor",
        "operationId": "get_vendor_api_v1_vendors__vendor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "vendors"
        ],
        "summary": "Update Vendor",
        "operationId": "update_vendor_api_v1_vendors__vendor_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vendor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vendors/{vendor_id}/status": {
      "post": {
        "tags": [
          "vendors"
        ],
        "summary": "Change Vendor Status",
        "description": "Approve, suspend or block a vendor. Suspended/blocked vendors are rejected by policy rule VEN-001.",
        "operationId": "change_vendor_status_api_v1_vendors__vendor_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vendor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorStatusChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContractStatus": {
        "type": "string",
        "enum": [
          "none",
          "active",
          "expired"
        ],
        "title": "ContractStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpendCategory": {
        "type": "string",
        "enum": [
          "it_services",
          "travel",
          "facilities",
          "marketing",
          "professional_services",
          "hardware",
          "software",
          "logistics",
          "office_supplies",
          "manufacturing"
        ],
        "title": "SpendCategory",
        "description": "Top-level spend taxonomy. Also the label set for the spend-classification model (P9)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VendorCreate": {
        "properties": {
          "legal_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Name"
          },
          "corporate_number": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{13}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corporate Number"
          },
          "invoice_registration_number": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^T\\d{13}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Registration Number"
          },
          "risk_level": {
            "$ref": "#/components/schemas/VendorRiskLevel",
            "default": "unknown"
          },
          "contract_status": {
            "$ref": "#/components/schemas/ContractStatus",
            "default": "none"
          },
          "contract_expires_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Expires On"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "country": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "title": "Country",
            "default": "JP"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "default": "JPY"
          },
          "payment_terms_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 0.0,
            "title": "Payment Terms Days",
            "default": 30
          },
          "website": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/SpendCategory"
            },
            "type": "array",
            "minItems": 1,
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "name",
          "categories"
        ],
        "title": "VendorCreate"
      },
      "VendorOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "legal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Name"
          },
          "corporate_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corporate Number"
          },
          "invoice_registration_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Registration Number"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "contract_status": {
            "type": "string",
            "title": "Contract Status"
          },
          "contract_expires_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Expires On"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "payment_terms_days": {
            "type": "integer",
            "title": "Payment Terms Days"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status_changed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Changed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "legal_name",
          "corporate_number",
          "invoice_registration_number",
          "categories",
          "status",
          "risk_level",
          "contract_status",
          "contract_expires_on",
          "rating",
          "contact_name",
          "contact_email",
          "phone",
          "country",
          "currency",
          "payment_terms_days",
          "website",
          "notes",
          "status_changed_at",
          "created_at",
          "updated_at"
        ],
        "title": "VendorOut"
      },
      "VendorPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VendorOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "VendorPage"
      },
      "VendorRiskLevel": {
        "type": "string",
        "enum": [
          "unknown",
          "low",
          "medium",
          "high"
        ],
        "title": "VendorRiskLevel"
      },
      "VendorStatus": {
        "type": "string",
        "enum": [
          "pending_review",
          "approved",
          "suspended",
          "blocked"
        ],
        "title": "VendorStatus"
      },
      "VendorStatusChange": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/VendorStatus"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "VendorStatusChange"
      },
      "VendorUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "legal_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Name"
          },
          "corporate_number": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{13}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corporate Number"
          },
          "invoice_registration_number": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^T\\d{13}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Registration Number"
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SpendCategory"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories"
          },
          "risk_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VendorRiskLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract_expires_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Expires On"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "payment_terms_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Terms Days"
          },
          "website": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "VendorUpdate"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /auth/login",
        "scheme": "bearer"
      }
    }
  }
}